+++ src/renderer/i18n/index.js
@@ -1,14 +1,24 @@
 export default function loadLanguage() {
     const context = require.context("./languages", false, /([a-z_]+)\.js$/i)
+    const context2 = require.context("element-ui/lib/locale/lang", false, /([a-z_]+)\.js$/i)
 
     const languages = context
         .keys()
-        .map((key) => ({ key, name: key.match(/([a-z_]+)\.js$/i)[1] }))
+        .map((key) => ({ key, name: key.match(/([a-z_-]+)\.js$/i)[1] }))
         .reduce(
-            (languages, {key, name}) => ({
-                ...languages,
-                [name]: context(key).lang
-            }),
+            (languages, {key, name}) => {
+                console.log(key,name)
+                let lang;
+                try {
+                    lang = Object.assign(context(key).lang, context2(key).default);
+                } catch(err) {
+                    lang = context(key).lang
+                }
+                return {
+                    ...languages,
+                    [name]: lang
+                }
+            },
             {}
         )