+++ src/renderer/main.js
@@ -25,14 +25,16 @@ if (!process.env.IS_WEB) {
   Vue.prototype.$ipcApi = require('./utils/ipcRenderer').default
 }
 
-Vue.use(ElementUI)
-
 // 创建 i18n
 Vue.use(VueI18n) // 新版本必须要这个，不知道为什么
 const i18n = new VueI18n({
-  locale: 'zh_CN', // 设置默认语言
+  locale: 'zh-CN', // 设置默认语言
   messages: languages, // 设置语言包
 });
+
+Vue.use(ElementUI, {
+  i18n: (key, value) => i18n.t(key, value)
+})
 
 Vue.config.productionTip = false
 /* eslint-disable no-new */